+++ packages/ozone/src/lexicon/types/tools/ozone/moderation/queryEvents.ts
@@ -47,6 +47,8 @@ export type QueryParams = {
   removedTags?: string[]
   reportTypes?: string[]
   policies?: string[]
+  /** If specified, only events where the modTool name matches any of the given values are returned */
+  modTool?: string[]
   cursor?: string
 }
 export type InputSchema = undefined